feat: api Вынести интерфейс Storager из конктерных апи
